requirementss.txt
absl-py==2.1.0
aext-assistant
aext-assistant-server
aext-core
aext-core-server
aext-panels
aext-panels-server
aext-share-notebook
aext-share-notebook-server
aext-shared
aiobotocore
aiohappyeyeballs
aiohttp
aioitertools
aiosignal
alabaster
altair
anaconda-anon-usage
anaconda-catalogs
anaconda-client
anaconda-cloud-auth
anaconda-navigator
anaconda-project
annotated-types
anyio
appdirs==1.4.4
archspec
argon2-cffi
argon2-cffi-bindings
arrow
astroid
astropy
astropy-iers-data
asttokens
astunparse==1.6.3
async-lru
atomicwrites==1.4.0
attrs
Automat
autopep8
Babel
backcall==0.2.0
bcrypt==4.1.2
beautifulsoup4
binaryornot
black
bleach
blinker
bokeh
boltons
botocore
Bottleneck
Brotli
cachetools
certifi
cffi
chardet
charset-normalizer
click
cloudpickle
colorama
colorcet
comm
conda
conda-build
conda-content-trust
conda-libmamba-solver
conda-pack
conda-package-handling
conda-repo-cli
conda-token
conda_index
conda_package_streaming
constantly
contourpy
cookiecutter
cryptography
cssselect
cycler
cytoolz
dask
dask-expr
datashader
debugpy
decorator
defusedxml
diff-match-patch
dill
distributed
distro
docopt==0.6.2
docstring-to-markdown
docutils
et-xmlfile
executing
fastjsonschema
filelock
flake8
Flask
Flask-Bcrypt==1.0.1
flask-cors==5.0.1
Flask-Mail==0.10.0
flask-marshmallow==1.3.0
Flask-MySQL==1.6.0
Flask-MySQLdb==2.0.0
Flask-SQLAlchemy==3.1.1
flatbuffers==24.3.25
fonttools
frozendict
frozenlist
fsspec
gast==0.6.0
gensim
gitdb
GitPython
google-pasta==0.2.0
greenlet
grpcio==1.67.1
h11
h5py
HeapDict
holoviews
httpcore
httpx
hvplot
hyperlink
idna
imagecodecs
imageio
imagesize
imbalanced-learn
importlib-metadata
incremental
inflection
iniconfig
intake
intervaltree
ipykernel
ipython==8.12.3
ipython-genutils
ipywidgets
isort
itemadapter
itemloaders
itsdangerous
jaraco.classes
jedi
jellyfish
Jinja2
jmespath
joblib
json5
jsonpatch
jsonpointer==2.1
jsonschema
jsonschema-specifications
jupyter
jupyter-console
jupyter-events
jupyter-lsp
jupyter_client
jupyter_core
jupyter_server
jupyter_server_terminals
jupyterlab
jupyterlab-pygments
jupyterlab-widgets
jupyterlab_server
keras==3.6.0
keyring
kiwisolver
lazy-object-proxy
lazy_loader
lckr_jupyterlab_variableinspector
libarchive-c
libclang==18.1.1
libmambapy
linkify-it-py
llvmlite
lmdb
locket
lxml==5.2.1
lz4
Markdown
markdown-it-py
MarkupSafe
marshmallow==3.26.1
matplotlib==3.9.2
matplotlib-inline
mccabe
mdit-py-plugins
mdurl
menuinst
mistune
mkl-service==2.4.0
mkl_fft
mkl_random
ml_dtypes==0.5.1
more-itertools
mpmath
msgpack
multidict
multipledispatch
mypy
mypy-extensions
mysql-connector-python==8.0.31
mysql-connector-repackaged==0.3.1
mysqlclient==2.2.7
namex==0.0.8
navigator-updater
nbclient
nbconvert
nbformat
nest-asyncio
networkx
nltk
notebook
notebook_shim
numba
numexpr
numpy==1.26.4
numpydoc
opencv-python==4.10.0.84
openpyxl
opt_einsum==3.4.0
optree==0.13.1
overrides
packaging
pandas==2.2.2
pandocfilters
panel
param
paramiko
parsel
parso
partd
pathspec
patsy
pexpect
pickleshare
pillow
pipreqs==0.5.0
pkce
pkginfo
platformdirs
plotly
pluggy
ply
prometheus-client
prompt-toolkit
Protego
protobuf==3.20.1
psutil
ptyprocess==0.7.0
pure-eval
py-cpuinfo
pyarrow
pyasn1
pyasn1-modules==0.2.8
pycodestyle
pycosat
pycparser
pyct
pycurl
pydantic
pydantic_core
pydeck
PyDispatcher
pydocstyle
pyerfa
pyflakes
Pygments
PyJWT
pylint
pylint-venv
pyls-spyder==0.4.0
PyMySQL==1.1.1
PyNaCl
pyodbc
pyOpenSSL
pyparsing
PyQt5==5.15.10
PyQt5-sip
PyQtWebEngine==5.15.6
PySocks
pytest
python-dateutil
python-dotenv
python-json-logger
python-lsp-black
python-lsp-jsonrpc
python-lsp-server
python-slugify
pytoolconfig
pytz
pyviz_comms
PyWavelets
pywin32==305.1
pywin32-ctypes
pywinpty==2.0.10
PyYAML
pyzmq
QDarkStyle
qstylizer==0.2.2
QtAwesome
qtconsole
QtPy
queuelib
referencing
regex
requests
requests-file
requests-toolbelt
rfc3339-validator
rfc3986-validator
rich
rope
rpds-py
Rtree
ruamel-yaml-conda
ruamel.yaml
ruamel.yaml.clib
s3fs
scikit-image
scikit-learn
scipy==1.13.1
Scrapy
seaborn
semver
Send2Trash
service-identity
setuptools==75.1.0
sip
six
smart-open
smmap
sniffio
snowballstemmer
sortedcontainers
soupsieve
Sphinx
sphinxcontrib-applehelp
sphinxcontrib-devhelp
sphinxcontrib-htmlhelp
sphinxcontrib-jsmath
sphinxcontrib-qthelp
sphinxcontrib-serializinghtml
spyder
spyder-kernels
SQLAlchemy
stack-data
statsmodels
streamlit
sympy
tables
tabulate
tblib
tenacity
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.0
tensorflow_intel==2.18.0
termcolor==2.5.0
terminado
text-unidecode
textdistance
tf_keras==2.19.0
threadpoolctl
three-merge
tifffile
tinycss2
tldextract
toml
tomli
tomlkit
toolz
tornado
tqdm
traitlets
truststore
Twisted
twisted-iocpsupport
typing_extensions
tzdata
uc-micro-py
ujson
unicodedata2
Unidecode
urllib3
w3lib
watchdog
wcwidth
webencodings
websocket-client
Werkzeug==3.1.3
whatthepatch
wheel==0.44.0
widgetsnbextension
win-inet-pton
wrapt
xarray
xlwings
xyzservices
yapf
yarg==0.1.9
yarl
zict
zipp
zope.interface
zstandard